move app.js and app.css to components
