+++ src/components/App/App.js
@@ -1,4 +1,4 @@
-import logo from './logo.svg';
+import logo from '../../../src/logo.svg';
 import './App.css';
 
 function App() {
